jigsaw-diffusion/pyproject.toml
[tool.poetry]
name = "jigsaw-diffusion"
version = "0.1.0"
description = "The diffusion model solving Jigsaw puzzles based on OpenAI's guided-diffusion"
authors = ["aeon <jerry73204@gmail.com>"]
readme = "README.md"
packages = [{include = "jigsaw_diffusion"}]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.0.0"
torchvision = "^0.15.1"
blobfile = "^2.0.2"
mpi4py = "^3.1.4"
json5 = "^0.9.14"
opencv-python = "^4.7.0.72"
pyserde = "^0.10.6"


[tool.poetry.group.test]  # This part can be left out

[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"
pytest-mock = "*"

[tool.poetry.scripts]
train = "jigsaw_diffusion.train:__main__"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
